+++ test/widgets/player_menu_anchor_test.dart
@@ -39,7 +39,6 @@ void main() {
               onSelectEmbedded: (_) {},
               onToggleOff: () {},
               onOpenSyncBar: () {},
-              onOpenStyleBar: () {},
               onAutoPick: () {},
               onClose: () {},
             ),